+++ fhe_http_core/src/fhe_traits/serializable/mod.rs
@@ -1,10 +1,10 @@
-pub mod fhe_value_serialable;
-pub mod key_serializable;
-pub mod value_serializable;
-pub mod zk_serializable;
+pub mod fhe_value;
+pub mod key;
+pub mod value;
+pub mod zk;
 
 // using prelude to re-export traits
-pub use fhe_value_serialable::FheValueSerializable;
-pub use key_serializable::KeySerializable;
-pub use value_serializable::ValueSerializable;
-pub use zk_serializable::ZkSerializable;
+pub use fhe_value::FheValueSerializable;
+pub use key::KeySerializable;
+pub use value::ValueSerializable;
+pub use zk::ZkSerializable;
